move ExampleStringUtils to bizfw package
