--- src/main/java/org/docksidestage/app/utils/ExampleStringUtils.java
@@ -1,17 +1,0 @@
-package org.docksidestage.app.utils;
-
-import org.springframework.lang.Nullable;
-
-/**
- * @author subaru
- * @author y.shimizu
- */
-public class ExampleStringUtils {
-    public static boolean isNotEmpty(@Nullable String str) {
-        return str != null && !str.isEmpty();
-    }
-
-    public static boolean isEmpty(@Nullable String str) {
-        return str == null || str.isEmpty();
-    }
-}
